+++ crates/tad-gui/src/compiler_thread.rs
@@ -8,7 +8,7 @@ use crate::names::NameGetter;
 use crate::sample_analyser::{self, SampleAnalysis};
 use crate::GuiMessage;
 
-use crate::audio_thread::AudioMessage;
+use crate::audio_thread::{AudioMessage, ChannelsMask};
 
 use std::collections::hash_map::Entry;
 use std::collections::HashMap;
@@ -101,7 +101,7 @@ pub enum ToCompiler {
 
     SongTabClosed(ItemId),
     SongChanged(ItemId, String),
-    CompileAndPlaySong(ItemId, String, Option<TickCounter>),
+    CompileAndPlaySong(ItemId, String, Option<TickCounter>, ChannelsMask),
     PlayInstrument(ItemId, PlaySampleArgs),
     PlaySample(ItemId, PlaySampleArgs),
 
@@ -1175,7 +1175,12 @@ fn bg_thread(
             ToCompiler::PlaySoundEffect(id) => {
                 if let Some(Some(sfx_data)) = sound_effects.get_output_for_id(&id) {
                     let song_data = Arc::new(sound_effect_to_song(sfx_data));
-                    sender.send_audio(AudioMessage::PlaySong(id, song_data, None));
+                    sender.send_audio(AudioMessage::PlaySong(
+                        id,
+                        song_data,
+                        None,
+                        ChannelsMask::ALL,
+                    ));
                 }
             }
 
@@ -1185,11 +1190,16 @@ fn bg_thread(
             ToCompiler::SongChanged(id, mml) => {
                 songs.edit_and_compile_song(id, mml, &pf_songs, &song_dependencies, &sender);
             }
-            ToCompiler::CompileAndPlaySong(id, mml, ticks_to_skip) => {
+            ToCompiler::CompileAndPlaySong(id, mml, ticks_to_skip, channels_mask) => {
                 sender.send_audio(AudioMessage::Pause);
                 songs.edit_and_compile_song(id, mml, &pf_songs, &song_dependencies, &sender);
                 if let Some(song) = songs.get_song_data(&id) {
-                    sender.send_audio(AudioMessage::PlaySong(id, song.clone(), ticks_to_skip));
+                    sender.send_audio(AudioMessage::PlaySong(
+                        id,
+                        song.clone(),
+                        ticks_to_skip,
+                        channels_mask,
+                    ));
                 }
             }
             ToCompiler::PlayInstrument(id, args) => {